Add files via upload
This report gives a brief overview of ethics in AI. It explains Ethics in Artificial Intelligence (AI), and its importance. This report describes how ethics play an important role in developing user-friendly AI. It covers the current development of new artificially intelligent technology such as Self-driving automobiles, nuclear power, a microchip in the brain, and the metaverse. Artificial intelligence that performs and acts ethically is known as ethical AI. To design acceptable ethical principles, norms, guidelines, laws, and regulations for AI, it is necessary to first recognize and analyze the ethical and moral challenges that AI may raise. AI has the ability to significantly disrupt our way of life and social norms. The window of opportunity to understand the consequences of these technologies and to detect their harmful consequences. In order to manage this new AI revolution, humanity must be proactive rather than reactive.